Move AttestationAgent code to its own package
This make it easier to build a subset of the code in google3. It
also changes the external API of this package to just be an
interface. This lets us avoid changing our unit tests while also
avoiding code duplication.
