+++ launcher/agent/agent_test.go
@@ -1,4 +1,4 @@
-package main
+package agent
 
 import (
 	"bytes"
